% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_check.r
\name{aa_check}
\alias{aa_check}
\alias{aa_check.DNAseq}
\title{Translate the sequence and check log likelihood as a final double check.}
\usage{
aa_check(x, ...)

\method{aa_check}{DNAseq}(x, ..., trans_table = 0, frame_offset = 0)
}
\arguments{
\item{x}{a ccs_reads class object.}

\item{...}{additional arguments to be passed between methods.}

\item{trans_table}{The translation table to use for translating from nucleotides to amino acids. Default is "auto", meaning
that the translation table will be inferred from the ccs_reads object's order.}

\item{frame_offset}{The offset to the reading frame to be applied for translation. By default the offset
is zero, so the first character in the framed sequence is considered the first nucelotide of the first codon.
Passing frame_offset = 1 would offset the sequence by one and therefore make the second character in the
framed sequence the the first nucelotide of the first codon.}
}
\value{
a class object of code{"ccs_reads"}
}
\description{
Translate the sequence and check log likelihood as a final double check.
}
\examples{
#previously called
ex_data = DNAseq(example_nt_string, name = 'ex1')
ex_data =  frame(ex_data)
ex_data = adjust(ex_data)
#run the aa check on the adjusted sequences
ex_data = aa_check(ex_data)
}
\seealso{
\code{\link{DNAseq}}

\code{\link{frame}}

\code{\link{adjust}}
}
