% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmms.r
\docType{data}
\name{aa_phmms}
\alias{aa_phmms}
\title{By order amino acid profile hidden markov models for coi5p.
Also contains the generic AA profile - 'unknown'}
\format{An object of class \code{list} of length 260.}
\usage{
aa_phmms
}
\description{
By order amino acid profile hidden markov models for coi5p.
Also contains the generic AA profile - 'unknown'
}
\keyword{internal}
