% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_and_adjust.r
\name{adj_seq}
\alias{adj_seq}
\title{Adjust the DNA sequence based on the ntPHMM path}
\usage{
adj_seq(frame_dat, path_out, censor_length = 3)
}
\arguments{
\item{frame_dat}{The read's framing data - generated by set_frame()}

\item{path_out}{The nucleotide PHMM path for the sequence.}

\item{censor_length}{number of base pairs in either direction of a PHMM correction
to convert to placeholder characters}
}
\description{
Adjust the DNA sequence based on the ntPHMM path
}
\keyword{internal}
