% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_and_adjust.r
\name{adjust}
\alias{adjust}
\alias{adjust.ccs_reads}
\title{Adjust the sequences based on the nt path outputs.}
\usage{
adjust(x, ...)

\method{adjust}{ccs_reads}(x, ..., censor_length = 3)
}
\arguments{
\item{x}{a ccs_reads class object.}

\item{...}{additional arguments to be passed between methods.}

\item{censor_length}{the number of base pairs in either direction of a PHMM correction
to convert to placeholder characters. Default is 3.}
}
\value{
a class object of code{"ccs_reads"}
}
\description{
Adjust the sequences based on the nt path outputs.
}
\examples{
#' #previously called
ex_data = build_ccs(ex_ccs_read_list, order = 'Diptera', id = 'SSGBC787-14')
ex_data =  frame(ex_data)
#adjust the sequence with default censor length is 3
ex_data = adjust(ex_data)
#with a custom censorship size
ex_data = adjust(ex_data, censor_length = 5)
}
\seealso{
\code{\link{build_ccs}}

\code{\link{frame}}
}
