% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_reads.r
\name{build_ccs}
\alias{build_ccs}
\title{Build a ccs_reads object from a DNA sequence string.}
\usage{
build_ccs(x, id = character(), order = "unknown")
}
\arguments{
\item{x}{a vector of nucleotide strings.}

\item{id}{and optional identification string.}

\item{order}{the taxonomic order corresponding to the sample. 
If taxonomy is not known, use the default 'unknown' option.}
}
\value{
an object of class code{"coi5p"}
}
\description{
Build a ccs_reads object from a DNA sequence string.
}
\examples{
#known taxonomy:
ex_data = build_ccs(ex_ccs_read_list, order = 'Diptera', id = 'custom_id')
#unknown taxonomy:
ex_data = build_ccs(ex_ccs_read_list,  id = 'custom_id')
}
