% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_seq.r
\name{col_mode}
\alias{col_mode}
\title{Determining the most common character in each column.}
\usage{
col_mode(col_data)
}
\value{
A single character string.
}
\description{
Determining the most common character in each column.
}
\keyword{internal}
