% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_seq.r
\name{consensus}
\alias{consensus}
\alias{consensus_sequence}
\title{Take the list of adjusted sequences and obtain the consensus sequence.}
\usage{
consensus_sequence(x)
}
\arguments{
\item{x}{The list of adjusted DNA sequences, post censorship and AA correction}
}
\description{
Take the list of adjusted sequences and obtain the consensus sequence.
}
\examples{

#build another sequence with an error
err_2= gsub('ccggatgaactgtatat', 'ccggatgaaactgtatat',example_nt_string)
ex_list = list( example_nt_string_errors,  err_2)
ex_out = denoise_list(ex_list)

barcode_seq = consensus_sequence(ex_out)
}
\seealso{
\code{\link{denoise_list}}
}
