% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.r
\name{consensus_sequence}
\alias{consensus_sequence}
\title{Take the list of adjusted sequences and obtain the consensus sequence.}
\usage{
consensus_sequence(x)
}
\arguments{
\item{x}{The list of adjusted DNA sequences, post censorship and AA correction}
}
\description{
Take the list of adjusted sequences and obtain the consensus sequence.
}
\keyword{internal}
