% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{denoise_file}
\alias{denoise_file}
\alias{denoise_file.default}
\title{Denoise the sequence data from a given file.}
\usage{
denoise_file(x, ...)

\method{denoise_file}{default}(x, ..., file_type = "fastq",
  multicore = FALSE)
}
\arguments{
\item{x}{The name of the file to denoise sequences from.}

\item{...}{additional arguments to be passed to the \link{denoise} function.}

\item{file_type}{The format of the file to be denoised. Options are fastq or fasta. Default is fastq.}

\item{multicore}{An integer specifying the number of cores over which to multithread the denoising process. 
Default is FALSE, meaning the process is not multithreaded.}
}
\description{
Denoise the sequence data from a given file.
}
\examples{
fastq_dat_file = system.file('extdata/ccs_subset.fastq', package = 'coiDenoiser')
denoise_file(fastq_dat_file, to_file = FALSE)

}
\seealso{
\code{\link{denoise}}
}
