% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmms.r
\docType{data}
\name{ex_ccs_read_list}
\alias{ex_ccs_read_list}
\title{A list of coi5p DNA barcode sequences used in the examples provided
within the package documentation}
\format{An object of class \code{list} of length 5.}
\usage{
ex_ccs_read_list
}
\description{
A list of coi5p DNA barcode sequences used in the examples provided
within the package documentation
}
\keyword{datasets}
