% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmms.r
\docType{data}
\name{ex_nt_list}
\alias{ex_nt_list}
\title{An example of a list of four coi5p sequences, each containing indel errors.}
\format{An object of class \code{list} of length 4.}
\usage{
ex_nt_list
}
\description{
An example of a list of four coi5p sequences, each containing indel errors.
}
\keyword{datasets}
