% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmms.r
\docType{data}
\name{example_nt_string_errors}
\alias{example_nt_string_errors}
\title{Example coi5p DNA sequence string with insertion and deletion errors.}
\format{An object of class \code{character} of length 1.}
\usage{
example_nt_string_errors
}
\description{
This string of barcode data is used in the package documentation's examples
and within the vignette demonstrating how to use the package.
}
\keyword{datasets}
