% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_check.r
\name{first_stop}
\alias{first_stop}
\title{Find the first stop codon in the amino acid sequence. 
Return the corresponding nucleotide mask start position.}
\usage{
first_stop(aa_str)
}
\description{
Find the first stop codon in the amino acid sequence. 
Return the corresponding nucleotide mask start position.
}
\keyword{internal}
