% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_reads.r
\name{new_ccs_reads}
\alias{new_ccs_reads}
\title{Build ccs_reads class instance.}
\usage{
new_ccs_reads(x = c(), id = character(), order = "UNK")
}
\description{
Build ccs_reads class instance.
}
\keyword{internal}
