% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAseq.r
\name{new_DNAseq}
\alias{new_DNAseq}
\title{Build a new DNAseq class instance.}
\usage{
new_DNAseq(x = character(), name = character(), phred = NULL)
}
\description{
Build a new DNAseq class instance.
}
\keyword{internal}
