% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmms.r
\docType{data}
\name{nt_phmms}
\alias{nt_phmms}
\title{By order nucleotide profile hidden markov models for coi5p.
Also contains the generic nucleotide profile - 'unknown'}
\format{An object of class \code{list} of length 260.}
\usage{
nt_phmms
}
\description{
By order nucleotide profile hidden markov models for coi5p.
Also contains the generic nucleotide profile - 'unknown'
}
\keyword{internal}
