% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_seq.r
\name{outseq}
\alias{outseq}
\alias{outseq.DNAseq}
\title{Get the final denoised output sequence for a read.}
\usage{
outseq(x, ...)

\method{outseq}{DNAseq}(x, keep_flanks = TRUE, ambig_char = "N",
  adjust_limit = 5)
}
\arguments{
\item{x}{a DNAseq class object.}

\item{...}{additional arguments to be passed between methods.}

\item{keep_flanks}{Default is TRUE.}

\item{ambig_char}{The character to use for ambigious positions in the sequence.}

\item{adjust_limit}{the maximum number of corrections that can be applied to a sequence read. If this number is exceeded 
then the entire read is masked with ambigious characters. Default is 5.}
}
\description{
Get the final denoised output sequence for a read.
}
