% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.ccs_reads}
\alias{print.ccs_reads}
\title{Print a concise summary of the ccs_reads object}
\usage{
\method{print}{ccs_reads}(x, ...)
}
\description{
Print a concise summary of the ccs_reads object
}
\keyword{internal}
