% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{read_fasta}
\alias{read_fasta}
\title{Read in raw data from a fasta file.}
\usage{
read_fasta(x)
}
\arguments{
\item{x}{The name of the fasta file to read data from}
}
\description{
Read in raw data from a fasta file.
}
\examples{
filename = system.file('extdata/ccs_subset.fasta', package = 'coiDenoiser')
data = read_fasta(filename)
}
