% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_seq.r
\name{set_frame}
\alias{set_frame}
\title{Take an input sequence and get it into the reading frame.
Uses the path of the sequence through the PHMM to located
the first long contigious stretch of 1s (matches)
Note - there is an imbalance in operation, namely the function
checks for a run of 10 additional 1s on the front of the sequence, and for
a run of only 3 additional matches on the back. The front of the sequence is high
priority because it is necessary to ensure the sequence is in frame.
whereas matching on the back is more lenient as insertions and deletions
can be tolerated here without large implications for the rest of the sequence}
\usage{
set_frame(org_seq_vec, path_out)
}
\description{
Take an input sequence and get it into the reading frame.
Uses the path of the sequence through the PHMM to located
the first long contigious stretch of 1s (matches)
Note - there is an imbalance in operation, namely the function
checks for a run of 10 additional 1s on the front of the sequence, and for
a run of only 3 additional matches on the back. The front of the sequence is high
priority because it is necessary to ensure the sequence is in frame.
whereas matching on the back is more lenient as insertions and deletions
can be tolerated here without large implications for the rest of the sequence
}
\keyword{internal}
