% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phmms.r
\docType{data}
\name{trans_df}
\alias{trans_df}
\title{Data frame containing the translation table recommendations.}
\format{An object of class \code{data.frame} with 4616 rows and 3 columns.}
\usage{
trans_df
}
\description{
Data frame containing the translation table recommendations.
}
\keyword{datasets}
