% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_seq.r
\name{triple_dels}
\alias{triple_dels}
\title{Look for triple deletes in the PHMM path.}
\usage{
triple_dels(x, path_start, path_end)
}
\description{
Look for triple deletes in the PHMM path.
}
\keyword{internal}
