% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_seq.r
\name{triple_ins}
\alias{triple_ins}
\title{Look for triple inserts in the PHMM path.}
\usage{
triple_ins(x, path_start, path_end)
}
\description{
Look for triple inserts in the PHMM path.
}
\keyword{internal}
