% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_reads.r
\name{validate_ccs_reads}
\alias{validate_ccs_reads}
\title{Validate the new ccs_reads class instance.}
\usage{
validate_ccs_reads(new_instance)
}
\description{
Validate the new ccs_reads class instance.
}
\keyword{internal}
