% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{write_fasta}
\alias{write_fasta}
\alias{write_fasta.ccs_reads}
\title{Output the denoised consensus sequence to a fasta file.}
\usage{
write_fasta(x, ...)

\method{write_fasta}{ccs_reads}(x, ..., filename = "denoised.fasta",
  append = TRUE)
}
\arguments{
\item{x}{a ccs_reads class object.}

\item{...}{additional arguments to be passed between methods.}

\item{filename}{The name of the file to output the data to. Default is "denoised.fasta".}

\item{append}{Should the ccs consensus sequence be appended to the output file?(TRUE) 
Or overwrite the file?(FALSE) Default is TRUE.}
}
\value{
a class object of code{"ccs_reads"}
}
\description{
Output the denoised consensus sequence to a fasta file.
}
\examples{
#previously called
ex_data = build_ccs(ex_ccs_read_list, order = 'Diptera', id = 'SSGBC787-14')
ex_data =  frame(ex_data)
ex_data = adjust(ex_data)
ex_data = consensus(ex_data)
#write to a fasta file with the default output file
write_fasta(ex_data)
#specify the path to a custom output file, overwrite its current contents
write_fasta(ex_data, filename = 'example_out.fasta', append = FALSE)
}
\seealso{
\code{\link{build_ccs}}

\code{\link{frame}}

\code{\link{adjust}}
}
