% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.r
\name{consensus}
\alias{consensus}
\alias{consensus.ccs_reads}
\title{Get the consensus sequence of the denoised circular consensus reads.}
\usage{
consensus(x, ...)

\method{consensus}{ccs_reads}(x, ..., ambig_char = "N")
}
\arguments{
\item{x}{a ccs_reads class object.}

\item{...}{additional arguments to be passed between methods.}

\item{ambig_char}{The character to use for ambigious positions in the sequence.}
}
\value{
a class object of code{"ccs_reads"}
}
\description{
This function takes the adjusted sequence reads and determines the consensus sequence.
Placeholder 'N' is added for instances where the nucleotide for the given sequence position 
cannot be stated with certainty.
}
\examples{
#previously run
ex_data = build_ccs(ex_ccs_read_list, order = 'Diptera', id = 'SSGBC787-14')
ex_data =  frame(ex_data)
ex_data = adjust(ex_data, censor_length = 3)
#get the consensus of the adjusted sequences.
ex_data = consensus(ex_data, ambig_char = 'N')
}
\seealso{
\code{\link{build_ccs}}

\code{\link{frame}}

\code{\link{adjust}}
}
