% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_seq.r
\name{consensus}
\alias{consensus}
\alias{consensus_sequence}
\title{Take the list of denoised sequences and obtain the consensus sequence.}
\usage{
consensus_sequence(x)
}
\arguments{
\item{x}{The list of adjusted DNA sequences, post censorship and AA correction}
}
\value{
A string containing the consensus nucleotide sequence.
}
\description{
Take the list of denoised sequences and obtain the consensus sequence.
}
\examples{
#denoise list of sequences with the k
ex_out = denoise_list(ex_nt_list, keep_flanks=FALSE)

barcode_seq = consensus_sequence(ex_out)
}
\seealso{
\code{\link{denoise_list}}
}
