% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{denoise}
\alias{denoise}
\title{Run the denoiser pipeline for a sequence read.}
\usage{
denoise(x, ..., name = character(), ambig_char = "N",
  censor_length = 5, to_file = TRUE, keep_flanks = TRUE,
  outformat = "fastq", filename = NULL, phred_placeholder = "#",
  aa_check = FALSE, trans_table = 0, frame_offset = 0,
  append = TRUE)
}
\arguments{
\item{x}{a DNA sequence string.}

\item{...}{additional arguments to be passed between methods.}

\item{name}{an optional character string. Identifier for the sequence.}

\item{ambig_char}{The character to use for ambigious positions in the sequence that is output to the file. Default is N.}

\item{censor_length}{the number of base pairs in either direction of a PHMM correction
to convert to placeholder characters. Default is 5.}

\item{to_file}{Boolean indicating whether the sequence should be written to a file. Default is TRUE.}

\item{keep_flanks}{Should the regions of the input sequence outside of the barcode region be readded to the denoised sequence
prior to outputting to the file. Default is TRUE. 
False will lead to only the denoised sequence for the 657bp barcode region being output to the file.}

\item{outformat}{The format of the output file. Options are fasta or fastq (default) format.}

\item{filename}{The name of the file to output the data to. Default filenames are respectively: denoised.fasta or denoised.fastq.}

\item{phred_placeholder}{The character to input for the phred score line. Default is '#'. Used with write_fastq only.}

\item{aa_check}{Boolean indicating whether the amino acid sequence should be generated and assessed for stop codons. Default = FALSE.}

\item{trans_table}{The translation table to use for translating from nucleotides to amino acids. Default is 0, meaning
that censored translation is performed (amigious codons ignored). Used only when aa_check = TRUE.}

\item{frame_offset}{The offset to the reading frame to be applied for translation. By default the offset
is zero, so the first character in the framed sequence is considered the first nucelotide of the first codon.
Passing frame_offset = 1 would offset the sequence by one and therefore make the second character in the
framed sequence the the first nucelotide of the first codon. Used only when aa_check = TRUE.}

\item{append}{Should the denoised sequence be appended to the output file?(TRUE) 
Or should the sequence overwrite the output file?(FALSE) Default is TRUE.}
}
\value{
a class object of code{"DNAseq"}
}
\description{
Run the denoiser pipeline for a sequence read.
}
\examples{
# Denoise example sequence with default paramaters.
ex_data = denoise(example_nt_string, name = 'example_sequence_1')
}
