% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_seq.r
\name{dir_check}
\alias{dir_check}
\title{Take an input sequence and align both the forward and reverse compliments to the PHMM}
\usage{
dir_check(x)
}
\arguments{
\item{x}{a DNAseq class object.}
}
\description{
The fuction returns the sequence, DNAbin and Path and score of the optimal orientation.
Optimal orientation is determined by the direction with the longer string of consecutive 
ones in the path
}
