% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_seq.r
\name{frame}
\alias{frame}
\alias{frame.DNAseq}
\title{Take a DNAseq object and put sequences in common reading frame}
\usage{
frame(x, ...)

\method{frame}{DNAseq}(x, ..., dir_check = TRUE, min_match = 100,
  terminate_rejects = TRUE)
}
\arguments{
\item{x}{a DNAseq class object.}

\item{...}{additional arguments to be passed between methods.}

\item{dir_check}{Should both the forward and reverse compliments be considered?}

\item{min_match}{The minimum number of sequential matches to the PHMM for a sequence to be denoised.
Otherwise flag the sequence as a reject.}

\item{terminate_rejects}{Should a check be made to enusre minimum homology of the input sequence to the PHMM.
Makes sure there are not more than 400 consecutive insert states and not less than the minimum number of
sequential matches. Default is True}
}
\value{
a class object of code{"DNAseq"}
}
\description{
Take a DNAseq object and put sequences in common reading frame
}
\examples{
#previously called

}
\seealso{
\code{\link{DNAseq}}
}
