% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_and_adjust.r
\name{frame}
\alias{frame}
\alias{frame.DNAseq}
\title{Take a DNAseq object and put sequences in common reading frame}
\usage{
frame(x, ...)

\method{frame}{DNAseq}(x, ...)
}
\arguments{
\item{x}{a DNAseq class object.}

\item{...}{additional arguments to be passed between methods.}
}
\value{
a class object of code{"DNAseq"}
}
\description{
Take a DNAseq object and put sequences in common reading frame
}
\examples{
#previously called
ex_data = DNAseq(ex_ccs_read_list, id = 'SSGBC787-14')
#frame the ccs reads, will utilize the PHMM corresponding to the order specified in the object.
ex_data =  frame(ex_data)
}
\seealso{
\code{\link{DNAseq}}
}
