% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{meta_check}
\alias{meta_check}
\title{Run the log file and reject processing code after running the denoise pipeline.}
\usage{
meta_check(x, log_data = list(), log_file = FALSE,
  keep_rejects = FALSE, reject_filename = "rejects.fastq", ...)
}
\description{
Run the log file and reject processing code after running the denoise pipeline.
}
\keyword{internal}
