% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_seq.r
\name{outseq}
\alias{outseq}
\alias{outseq.DNAseq}
\title{Get the final denoised output sequence for a read.}
\usage{
outseq(x, ...)

\method{outseq}{DNAseq}(x, keep_flanks = TRUE, ambig_char = "N",
  max_adjustments = 3)
}
\arguments{
\item{x}{a DNAseq class object.}

\item{...}{additional arguments to be passed between methods.}

\item{keep_flanks}{Default is TRUE.}

\item{ambig_char}{The character to use for ambigious positions in the sequence.}

\item{max_adjustments}{How many adjustments can be made to a read? If adjust exceeded this number
the entire read will be masked. Default is 3 corrections permitted.}
}
\description{
Get the final denoised output sequence for a read.
}
