% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred_check.r
\name{phred_check}
\alias{phred_check}
\alias{phred_check.DNAseq}
\title{Check the numeric phred scores and}
\usage{
phred_check(x, ...)

\method{phred_check}{DNAseq}(x, ..., min_avg_qv = 20,
  max_perc_low = 0.25, max_perc_ultra_low = 0.05)
}
\arguments{
\item{x}{a DNAseq class object.}

\item{max_perc_low}{The maximum percentage of nucleotides in the string with QV values lower than 20. Default is 25%}

\item{max_perc_ultra_low}{The maximum percentage of nucleotides in the string with QV values lower than 10. Default is 5%}
}
\description{
Check the numeric phred scores and
}
