% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{write_fasta}
\alias{write_fasta}
\alias{write_fasta.DNAseq}
\title{Output the denoised consensus sequence to a fasta file.}
\usage{
write_fasta(x, ...)

\method{write_fasta}{DNAseq}(x, ..., outfile = "denoised.fasta",
  append = TRUE)
}
\arguments{
\item{x}{a DNAseq class object.}

\item{...}{additional arguments to be passed between methods.}

\item{outfile}{The name of the file to output the data to. Default is "denoised.fasta".}

\item{append}{Should the ccs consensus sequence be appended to the output file?(TRUE) 
Or overwrite the file?(FALSE) Default is TRUE.}
}
\value{
a class object of code{"DNAseq"}
}
\description{
Output the denoised consensus sequence to a fasta file.
}
\seealso{
\code{\link{DNAseq}}

\code{\link{frame}}

\code{\link{adjust}}
}
